Перенес бизнес логику из view в VC, сделал отдельный container
