+++ Source/OmegaBank/Sources/Presentation/Map/MapViewController.swift
@@ -2,60 +2,68 @@
 //  MapViewController.swift
 //  OmegaBank
 //
-//  Created by Konsantin Makhov on 28.06.2021.
+//  Created by Konsantin Makhov on 19.07.2021.
 //  Copyright © 2021 RedMadRobot. All rights reserved.
 //
 
 import CoreLocation
 import MapKit
-import struct OmegaBankAPI.Office
 
-final class MapViewController: PageViewController, AlertPresentable {
+final class MapViewController: UIViewController, AlertPresentable {
     
-    // MARK: - Private properties
+    // MARK: - Private types
     
-    private let officesService: BankPlacesService
-    private var locationManager: CLLocationManager
-    private var progress: Progress?
+    private enum CameraZoom {
+        case zoomIn
+        case zoomOut
+    }
+    
+    // MARK: - Constants
+    
+    private enum Constants {
+        static let coordinateMoscow = CLLocationCoordinate2D(latitude: 55.751244, longitude: 37.618423)
+        static let scaleMoscow = 35000.0
+        static let scaleUserLocation = 1000.0
+        static let coordinateMoscowRegion = MKCoordinateRegion(
+            center: coordinateMoscow,
+            latitudinalMeters: scaleMoscow,
+            longitudinalMeters: scaleMoscow)
+        static let userSpan = MKCoordinateSpan(latitudeDelta: scaleUserLocation, longitudeDelta: scaleUserLocation)
+        
+        static let durationAnimation = 0.2
+    }
+    
+    // MARK: - Private Properties
+    
+    private let mapView = MapView()
+    private let locationManager: CLLocationManager
     private var locationStatus: CLAuthorizationStatus?
-    private var errorViewController: ErrorViewController?
-
-    private var mapView: MapView
     
     // MARK: - Init
     
-    init(officesService: BankPlacesService = ServiceLayer.shared.officesService) {
-        self.officesService = officesService
-        self.locationManager = CLLocationManager()
-        self.mapView = MapView(locationManager: locationManager)
-
-        super.init(title: "Map", tabBarImage: #imageLiteral(resourceName: "map"))
+    init(annotations: [MKAnnotation], locationManager: CLLocationManager = CLLocationManager()) {
+        self.locationManager = locationManager
+        mapView.addAnnotations(annotations)
         
-        navigationItem.title = "Offices & Bankomats"
+        super.init(nibName: nil, bundle: nil)
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    deinit {
-        progress?.cancel()
-    }
-    
-    // MARK: - View controller
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         mapView.delegate = self
+        mapView.mapControlsDelegate = self
         locationManager.delegate = self
         
         registerMapAnnotationView()
-        
-        loadOffices()
-        
         updateLocationStatus(CLLocationManager.authorizationStatus())
     }
+    
+    // MARK: - Public methods
     
     override func loadView() {
         view = mapView
@@ -69,6 +77,36 @@ final class MapViewController: PageViewController, AlertPresentable {
         mapView.registerAnnotationView(AtmMarkerAnnotationView.self)
     }
     
+    private func showMoscow() {
+        mapView.setRegion(Constants.coordinateMoscowRegion, animated: true)
+    }
+    
+    private func showUserLocation() {
+        guard let location = locationManager.location else { return }
+        
+        let region = MKCoordinateRegion(
+            center: location.coordinate,
+            latitudinalMeters: Constants.scaleUserLocation,
+            longitudinalMeters: Constants.scaleUserLocation)
+        
+        mapView.setRegion(region, animated: true)
+    }
+    
+    private func changeCamera(with zoom: CameraZoom) {
+        var region: MKCoordinateRegion = mapView.region
+        
+        switch zoom {
+        case .zoomOut:
+            region.span.latitudeDelta = min(region.span.latitudeDelta * 2.0, 180.0)
+            region.span.longitudeDelta = min(region.span.longitudeDelta * 2.0, 180.0)
+        case .zoomIn:
+            region.span.latitudeDelta /= 2.0
+            region.span.longitudeDelta /= 2.0
+        }
+        
+        mapView.animatedZoom(zoomRegion: region, duration: Constants.durationAnimation)
+    }
+    
     private func updateLocationStatus(_ status: CLAuthorizationStatus) {
         switch status {
         case .notDetermined:
@@ -77,16 +115,16 @@ final class MapViewController: PageViewController, AlertPresentable {
         case .restricted:
             mapView.locationButton.isHidden = true
             mapView.showsUserLocation = false
-            mapView.showMoscow()
+            showMoscow()
         case .denied:
             mapView.locationButton.isHidden = false
             mapView.showsUserLocation = false
-            mapView.showMoscow()
+            showMoscow()
         case .authorizedAlways, .authorizedWhenInUse:
             mapView.locationButton.isHidden = false
             locationManager.requestLocation()
             mapView.showsUserLocation = true
-            mapView.showUserLocation()
+            showUserLocation()
         default:
             break
         }
@@ -108,33 +146,6 @@ final class MapViewController: PageViewController, AlertPresentable {
                     UIApplication.shared.open(settingsUrl)
                 }
             ])
-    }
-    
-    private func showError(_ item: ErrorItem, onAction: VoidClosure? = nil) {
-        let vc = ErrorViewController(item, onAction: onAction)
-        addChildViewController(vc, to: view)
-        errorViewController = vc
-    }
-    
-    private func removeError() {
-        errorViewController?.removeChildFromParent()
-        errorViewController = nil
-    }
-    
-    /// Загрузка офисов
-    private func loadOffices() {
-        progress = officesService.load { [weak self] result in
-            switch result {
-            case .success(let (offices, atms)):
-                self?.mapView.addAnnotations(offices)
-                self?.mapView.addAnnotations(atms)
-            case .failure(let error):
-                self?.showError(.error(error), onAction: { [weak self] in
-                    self?.removeError()
-                    self?.loadOffices()
-                })
-            }
-        }
     }
 }
 
@@ -160,7 +171,7 @@ extension MapViewController: MKMapViewDelegate {
             view.clusteringIdentifier = String(describing: BankPointMarkerAnnotationView.self)
             view.setup(annotation)
             return view
-        } 
+        }
         return nil
     }
     
@@ -198,8 +209,25 @@ extension MapViewController: CLLocationManagerDelegate {
     }
     
     func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        mapView.showUserLocation()
+        showUserLocation()
     }
+}
+
+// MARK: - MapViewControlsDelegate
+
+extension MapViewController: MapViewControlsDelegate {
+    func mapViewControlsDidSelectLocationButton(_ mapView: MapView) {
+        showUserLocation()
+    }
+    
+    func mapViewControlsDidSelectZoomInButton(_ mapView: MapView) {
+        changeCamera(with: .zoomIn)
+    }
+    
+    func mapViewControlsDidSelectZoomOutButton(_ mapView: MapView) {
+        changeCamera(with: .zoomOut)
+    }
+    
 }
 
 // MARK: - MKMapView
